requirements.txt
python-telegram-bot
supabase
stripe
python-dotenv